requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
boto3==1.37.31
botocore==1.37.31
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
duckduckgo_search==8.0.0
elastic-transport==8.17.1
elasticsearch==8.18.0
fastapi==0.115.12
filetype==1.2.0
firecrawl-py==2.2.0
frozenlist==1.5.0
geojson==2.5.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
groq==0.22.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.23
langchain-anthropic==0.3.10
langchain-aws==0.2.18
langchain-community==0.3.21
langchain-core==0.3.51
langchain-elasticsearch==0.3.2
langchain-google-genai==2.0.11
langchain-groq==0.2.5
langchain-ollama==0.2.3
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langgraph==0.3.27
langgraph-checkpoint==2.0.24
langgraph-checkpoint-postgres==2.0.19
langgraph-checkpoint-sqlite==2.0.6
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langgraph-supervisor==0.0.16
langsmith==0.1.147
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.2
mypy-extensions==1.0.0
narwhals==1.34.1
nest-asyncio==1.6.0
numexpr==2.10.2
numpy==2.2.4
ollama==0.4.7
openai==1.72.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
primp==0.14.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic-settings==2.6.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.25.5
pyowm==3.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
s3transfer==0.11.4
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
streamlit==1.40.2
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.32.1
websockets==15.0.1
xxhash==3.5.0
yarl==1.19.0
